requirements.txt
imageio
opencv-python
matplotlib
tensorflow-macos
tensorflow-hub